python2/tests/test_requirements.txt
alibabacloud-rpc-util-py2
alibabacloud-credentials-py2
alibabacloud-tea-util-py2
alibabacloud-tea-py2
